refactor(core): group cimd modules under a dedicated directory
